#775 moved DefaultResourceManagerClassLoader to resource impl
